Clean up some header includes to slightly improve the build time (#1319)
* Move the `RuntimeParameter`s (which are only used in a few places but are very expensive to parse) into their own header. They used to be part of the global `Constants.h` file, which is included almost everywhere.
* Replace `gtest.h` by `gtest_prod.h` in places where only the `FRIEND_TEST` macro is needed.
* Move some code in the expression module from header files to `.cpp` files.
* Unify the syntax of several includes.
