+++ src/engine/sparqlExpressions/SparqlExpression.h
@@ -7,24 +7,13 @@
 
 #include <memory>
 #include <span>
-#include <variant>
 #include <vector>
 
-#include "engine/CallFixedSize.h"
-#include "engine/QueryExecutionContext.h"
-#include "engine/ResultTable.h"
-#include "engine/sparqlExpressions/SetOfIntervals.h"
 #include "engine/sparqlExpressions/SparqlExpressionPimpl.h"
 #include "engine/sparqlExpressions/SparqlExpressionTypes.h"
-#include "engine/sparqlExpressions/SparqlExpressionValueGetters.h"
-#include "global/Id.h"
 #include "parser/data/Variable.h"
-#include "util/ConstexprSmallString.h"
 
 namespace sparqlExpression {
-
-// TODO<joka921>  Move the definitions of the functions into a
-// `SparqlExpression.cpp`
 
 /// Virtual base class for an arbitrary Sparql Expression which holds the
 /// structure of the expression as well as the logic to evaluate this expression
@@ -43,123 +32,70 @@ class SparqlExpression {
 
   /// Return all variables and IRIs, needed for certain parser methods.
   /// TODO<joka921> should be called getStringLiteralsAndVariables
-  virtual vector<const Variable*> containedVariables() const final {
-    vector<const Variable*> result;
-    // Recursively aggregate the strings from all children.
-    for (const auto& child : children()) {
-      auto variablesFromChild = child->containedVariables();
-      result.insert(result.end(), variablesFromChild.begin(),
-                    variablesFromChild.end());
-    }
-
-    // Add the strings from this expression.
-    auto locallyAdded = getContainedVariablesNonRecursive();
-    for (auto& el : locallyAdded) {
-      result.push_back(&el);
-    }
-    return result;
-  }
+  virtual std::vector<const Variable*> containedVariables() const final;
 
   /// Return all the variables that occur in the expression, but are not
   /// aggregated.
-  virtual vector<Variable> getUnaggregatedVariables() {
-    // Default implementation: This expression adds no variables, but all
-    // unaggregated variables from the children remain unaggregated.
-    std::vector<Variable> result;
-    for (const auto& child : children()) {
-      auto childResult = child->getUnaggregatedVariables();
-      result.insert(result.end(), std::make_move_iterator(childResult.begin()),
-                    std::make_move_iterator(childResult.end()));
-    }
-    return result;
-  }
+  virtual std::vector<Variable> getUnaggregatedVariables();
 
   // Return true iff this expression contains an aggregate like SUM, COUNT etc.
   // This information is needed to check if there is an implicit GROUP BY in a
   // query because any of the selected aliases contains an aggregate.
-  virtual bool containsAggregate() const final {
-    if (isAggregate()) return true;
-    return std::ranges::any_of(children(), [](const Ptr& child) {
-      return child->containsAggregate();
-    });
-  }
+  virtual bool containsAggregate() const final;
 
   // Check if expression is aggregate
-  virtual bool isAggregate() const { return false; }
+  virtual bool isAggregate() const;
 
   // Check if an expression is distinct (only applies to aggregates)
-  virtual bool isDistinct() const {
-    AD_THROW(
-        "isDistinct() maybe only called for aggregate expressions. If this is "
-        "an aggregate expression, then the implementation of isDistinct() is "
-        "missing for this expression. Please report this to the developers.");
-  }
+  virtual bool isDistinct() const;
 
   // Replace child at index `childIndex` with `newExpression`
   virtual void replaceChild(size_t childIndex,
-                            std::unique_ptr<SparqlExpression> newExpression) {
-    AD_CONTRACT_CHECK(childIndex < children().size());
-    children()[childIndex] = std::move(newExpression);
-  }
+                            std::unique_ptr<SparqlExpression> newExpression);
 
   /// Get a unique identifier for this expression, used as cache key.
   virtual string getCacheKey(const VariableToColumnMap& varColMap) const = 0;
 
   /// Get a short, human readable identifier for this expression.
-  virtual const string& descriptor() const final { return _descriptor; }
-  virtual string& descriptor() final { return _descriptor; }
+  virtual const string& descriptor() const final;
+  virtual string& descriptor() final;
 
   /// For the pattern trick we need to know, whether this expression
   /// is a non-distinct count of a single variable. In this case we return
   /// the variable. Otherwise we return std::nullopt.
   virtual std::optional<SparqlExpressionPimpl::VariableAndDistinctness>
-  getVariableForCount() const {
-    return std::nullopt;
-  }
+  getVariableForCount() const;
 
   /// Helper function for getVariableForCount() : If this
   /// expression is a single variable, return the name of this variable.
   /// Otherwise, return std::nullopt.
-  virtual std::optional<::Variable> getVariableOrNullopt() const {
-    return std::nullopt;
-  }
+  virtual std::optional<::Variable> getVariableOrNullopt() const;
 
   // For the following three functions (`containsLangExpression`,
   // `getLanguageFilterExpression`, and `getEstimatesForFilterExpression`, see
   // the documentation of the functions of the same names in
   // `SparqlExpressionPimpl.h`. Each of them has a default implementation that
   // is correct for most of the expressions.
-  virtual bool containsLangExpression() const {
-    return std::ranges::any_of(children(),
-                               [](const SparqlExpression::Ptr& child) {
-                                 return child->containsLangExpression();
-                               });
-  }
+  virtual bool containsLangExpression() const;
 
   // ___________________________________________________________________________
   using LangFilterData = SparqlExpressionPimpl::LangFilterData;
-  virtual std::optional<LangFilterData> getLanguageFilterExpression() const {
-    return std::nullopt;
-  }
+  virtual std::optional<LangFilterData> getLanguageFilterExpression() const;
 
   // ___________________________________________________________________________
   using Estimates = SparqlExpressionPimpl::Estimates;
   virtual Estimates getEstimatesForFilterExpression(
       uint64_t inputSizeEstimate,
       [[maybe_unused]] const std::optional<Variable>& primarySortKeyVariable)
-      const {
-    // Default estimates: Each element can be computed in O(1) and nothing is
-    // filtered out.
-    return {inputSizeEstimate, inputSizeEstimate};
-  }
+      const;
 
   // Returns true iff this expression is a simple constant. Default
   // implementation returns `false`.
-  virtual bool isConstantExpression() const { return false; }
+  virtual bool isConstantExpression() const;
 
   // Returns true iff this expression is a STR(...) expression.  Default
   // implementation returns `false`.
-  virtual bool isStrExpression() const { return false; }
+  virtual bool isStrExpression() const;
 
   // __________________________________________________________________________
   virtual ~SparqlExpression() = default;
@@ -167,24 +103,13 @@ class SparqlExpression {
   // Returns all the children of this expression. Typically only used for
   // testing
   virtual std::span<const SparqlExpression::Ptr> childrenForTesting()
-      const final {
-    return children();
-  }
+      const final;
 
-  virtual std::vector<SparqlExpression::Ptr> moveChildrenOut() && final {
-    auto span = children();
-    return {std::make_move_iterator(span.begin()),
-            std::make_move_iterator(span.end())};
-  }
+  virtual std::vector<SparqlExpression::Ptr> moveChildrenOut() && final;
 
   // Get the direct child expressions.
-  virtual std::span<SparqlExpression::Ptr> children() final {
-    return childrenImpl();
-  }
-  virtual std::span<const SparqlExpression::Ptr> children() const final {
-    auto children = const_cast<SparqlExpression&>(*this).children();
-    return {children.data(), children.size()};
-  }
+  virtual std::span<SparqlExpression::Ptr> children() final;
+  virtual std::span<const SparqlExpression::Ptr> children() const final;
 
  private:
   virtual std::span<SparqlExpression::Ptr> childrenImpl() = 0;
@@ -192,29 +117,17 @@ class SparqlExpression {
   // Helper function for strings(). Get all variables, iris, and string literals
   // that are included in this expression directly, ignoring possible child
   // expressions.
-  virtual std::span<const Variable> getContainedVariablesNonRecursive() const {
-    // Default implementation: This expression adds no strings or variables.
-    return {};
-  }
+  virtual std::span<const Variable> getContainedVariablesNonRecursive() const;
 
  protected:
   // After calling this function, `isInsideAlias()` (see below) returns true for
   // this expression as well as for all its descendants. This function must be
   // called by all child classes that are aggregate expressions.
-  virtual void setIsInsideAggregate() final {
-    isInsideAggregate_ = true;
-    // Note: `child` is a `unique_ptr` to a non-const object. So we could
-    // technically use `const auto&` in the following loop, but this would be
-    // misleading (the pointer is used in a `const` manner, but the pointee is
-    // not).
-    for (auto& child : children()) {
-      child->setIsInsideAggregate();
-    }
-  }
+  virtual void setIsInsideAggregate() final;
 
   // Return true if this class or any of its ancestors in the expression tree is
   // an aggregate. For an example usage see the `LiteralExpression` class.
-  bool isInsideAggregate() const { return isInsideAggregate_; }
+  bool isInsideAggregate() const;
 };
 }  // namespace sparqlExpression
