+++ src/global/Constants.h
@@ -7,14 +7,13 @@
 
 #pragma once
 
+#include <chrono>
 #include <limits>
 #include <stdexcept>
 #include <string>
 #include <string_view>
 
 #include "util/MemorySize/MemorySize.h"
-#include "util/Parameters.h"
-#include "util/ParseableDuration.h"
 
 // For access to `memorySize` literals.
 using namespace ad_utility::memory_literals;
@@ -27,13 +26,8 @@ static constexpr ad_utility::MemorySize DEFAULT_MEM_FOR_QUERIES = 4_GB;
 static const size_t MAX_NOF_ROWS_IN_RESULT = 1'000'000;
 static const size_t MIN_WORD_PREFIX_SIZE = 4;
 static const char PREFIX_CHAR = '*';
-static const size_t MAX_NOF_NODES = 64;
-static const size_t MAX_NOF_FILTERS = 64;
 
-static const size_t BUFFER_SIZE_RELATION_SIZE = 1'000'000'000;
 static const size_t BUFFER_SIZE_DOCSFILE_LINE = 100'000'000;
-static const size_t DISTINCT_LHS_PER_BLOCK = 10'000;
-static const size_t USE_BLOCKS_INDEX_SIZE_TRESHOLD = 20'000;
 
 static const size_t TEXT_PREDICATE_CARDINALITY_ESTIMATE = 1'000'000'000;
 static const size_t TEXT_LIMIT_DEFAULT = std::numeric_limits<size_t>::max();
@@ -74,14 +68,9 @@ static const std::string INTERNAL_VARIABLE_PREFIX =
 constexpr std::string_view INTERNAL_BLANKNODE_VARIABLE_PREFIX =
     "?_QLever_internal_variable_bn_";
 
-static constexpr std::string_view TEXTSCORE_VARIABLE_PREFIX = "?ql_textscore_";
-static constexpr std::string_view ENTITY_VARIABLE_PREFIX = "?ql_entity_";
 static constexpr std::string_view SCORE_VARIABLE_PREFIX = "?ql_score_";
 static constexpr std::string_view MATCHINGWORD_VARIABLE_PREFIX =
     "?ql_matchingword_";
-
-// For anonymous nodes in Turtle.
-static const std::string ANON_NODE_PREFIX = "QLever-Anon-Node";
 
 static const std::string LANGUAGE_PREDICATE = makeInternalIri("langtag");
 
@@ -117,24 +106,16 @@ static const char XSD_UNSIGNED_INT_TYPE[] =
     "http://www.w3.org/2001/XMLSchema#unsignedInt";
 static const char XSD_UNSIGNED_SHORT_TYPE[] =
     "http://www.w3.org/2001/XMLSchema#unsignedShort";
-static const char XSD_UNSIGNED_BYTE_TYPE[] =
-    "http://www.w3.org/2001/XMLSchema#unsignedByte";
 static const char XSD_POSITIVE_INTEGER_TYPE[] =
     "http://www.w3.org/2001/XMLSchema#positiveInteger";
 constexpr inline char XSD_BOOLEAN_TYPE[] =
     "http://www.w3.org/2001/XMLSchema#boolean";
 static const char RDF_PREFIX[] = "http://www.w3.org/1999/02/22-rdf-syntax-ns#";
-static const char VALUE_DATE_TIME_SEPARATOR[] = "T";
-static const int DEFAULT_NOF_VALUE_INTEGER_DIGITS = 50;
-static const int DEFAULT_NOF_VALUE_EXPONENT_DIGITS = 20;
-static const int DEFAULT_NOF_VALUE_MANTISSA_DIGITS = 30;
-static const int DEFAULT_NOF_DATE_YEAR_DIGITS = 19;
 
 static const std::string INTERNAL_VOCAB_SUFFIX = ".vocabulary.internal";
 static const std::string EXTERNAL_VOCAB_SUFFIX = ".vocabulary.external";
 static const std::string MMAP_FILE_SUFFIX = ".meta";
 static const std::string CONFIGURATION_FILE = ".meta-data.json";
-static const std::string PREFIX_FILE = ".prefixes";
 
 static const std::string ERROR_IGNORE_CASE_UNSUPPORTED =
     "Key \"ignore-case\" is no longer supported. Please remove this key from "
@@ -178,10 +159,6 @@ static constexpr size_t PERCENTAGE_OF_TRIPLES_FOR_SORT_ESTIMATE = 5;
 // times this factor.
 static constexpr size_t MAKE_ROOM_SLACK_FACTOR = 2;
 
-// The version of the binary format of the pattern files. Has to be increased,
-// when this format is changed.
-static constexpr uint32_t PATTERNS_FILE_VERSION = 1;
-
 // The maximal number of columns an `IdTable` (an intermediate result of
 // query evaluation) may have to be able to use the more efficient `static`
 // implementation (For details see `IdTable.h`, `CallFixedSize.h` and the
@@ -201,49 +178,6 @@ constexpr std::chrono::milliseconds DESIRED_CANCELLATION_CHECK_INTERVAL{50};
 // the result, so the column order for PSO is S O PatternS PatternO.
 constexpr size_t ADDITIONAL_COLUMN_INDEX_SUBJECT_PATTERN = 3;
 constexpr size_t ADDITIONAL_COLUMN_INDEX_OBJECT_PATTERN = 4;
-
-inline auto& RuntimeParameters() {
-  using ad_utility::detail::parameterShortNames::Bool;
-  using ad_utility::detail::parameterShortNames::Double;
-  using ad_utility::detail::parameterShortNames::DurationParameter;
-  using ad_utility::detail::parameterShortNames::MemorySizeParameter;
-  using ad_utility::detail::parameterShortNames::SizeT;
-  // NOTE: It is important that the value of the static variable is created by
-  // an immediately invoked lambda, otherwise we get really strange segfaults on
-  // Clang 16 and 17.
-  // TODO<joka921> Figure out whether this is a bug in Clang or whether we
-  // clearly misunderstand something about static initialization.
-  static ad_utility::Parameters params = []() {
-    using namespace std::chrono_literals;
-    auto ensureStrictPositivity = [](auto&& parameter) {
-      parameter.setParameterConstraint(
-          [](std::chrono::seconds value, std::string_view parameterName) {
-            if (value <= 0s) {
-              throw std::runtime_error{absl::StrCat(
-                  "Parameter ", parameterName,
-                  " must be strictly positive, was ", value.count(), "s")};
-            }
-          });
-      return AD_FWD(parameter);
-    };
-    return ad_utility::Parameters{
-        // If the time estimate for a sort operation is larger by more than this
-        // factor than the remaining time, then the sort is canceled with a
-        // timeout exception.
-        Double<"sort-estimate-cancellation-factor">{3.0},
-        SizeT<"cache-max-num-entries">{1000},
-        MemorySizeParameter<"cache-max-size">{30_GB},
-        MemorySizeParameter<"cache-max-size-single-entry">{5_GB},
-        SizeT<"lazy-index-scan-queue-size">{20},
-        SizeT<"lazy-index-scan-num-threads">{10},
-        ensureStrictPositivity(
-            DurationParameter<std::chrono::seconds, "default-query-timeout">{
-                30s}),
-        SizeT<"lazy-index-scan-max-size-materialization">{1'000'000},
-        Bool<"group-by-hash-map-enabled">{false}};
-  }();
-  return params;
-}
 
 #ifdef _PARALLEL_SORT
 static constexpr bool USE_PARALLEL_SORT = true;
