+++ src/index/CompressedRelation.cpp
@@ -6,7 +6,9 @@
 
 #include <ranges>
 
+#include "engine/idTable/CompressedExternalIdTable.h"
 #include "engine/idTable/IdTable.h"
+#include "global/RuntimeParameters.h"
 #include "util/Cache.h"
 #include "util/CompressionUsingZstd/ZstdWrapper.h"
 #include "util/ConcurrentCache.h"
@@ -17,7 +19,6 @@
 #include "util/ThreadSafeQueue.h"
 #include "util/Timer.h"
 #include "util/TypeTraits.h"
-#include "util/jthread.h"
 
 using namespace std::chrono_literals;
