+++ src/index/IndexImpl.h
@@ -5,46 +5,40 @@
 //   2018-     Johannes Kalmbach (kalmbach@informatik.uni-freiburg.de)
 #pragma once
 
-#include <engine/ResultTable.h>
-#include <global/Pattern.h>
-#include <global/SpecialIds.h>
-#include <index/CompressedRelation.h>
-#include <index/ConstantsIndexBuilding.h>
-#include <index/DocsDB.h>
-#include <index/Index.h>
-#include <index/IndexBuilderTypes.h>
-#include <index/IndexMetaData.h>
-#include <index/PatternCreator.h>
-#include <index/Permutation.h>
-#include <index/StxxlSortFunctors.h>
-#include <index/TextMetaData.h>
-#include <index/Vocabulary.h>
-#include <index/VocabularyMerger.h>
-#include <parser/ContextFileParser.h>
-#include <parser/TripleComponent.h>
-#include <parser/TurtleParser.h>
-#include <util/BackgroundStxxlSorter.h>
-#include <util/BufferedVector.h>
-#include <util/CompressionUsingZstd/ZstdWrapper.h>
-#include <util/File.h>
-#include <util/Forward.h>
-#include <util/HashMap.h>
-#include <util/MmapVector.h>
-#include <util/json.h>
-
 #include <array>
-#include <fstream>
 #include <memory>
 #include <optional>
 #include <string>
-#include <stxxl/sorter>
-#include <stxxl/stream>
 #include <stxxl/vector>
 #include <vector>
 
+#include "engine/ResultTable.h"
 #include "engine/idTable/CompressedExternalIdTable.h"
+#include "global/Pattern.h"
+#include "global/SpecialIds.h"
+#include "index/CompressedRelation.h"
+#include "index/ConstantsIndexBuilding.h"
+#include "index/DocsDB.h"
+#include "index/Index.h"
+#include "index/IndexBuilderTypes.h"
+#include "index/IndexMetaData.h"
+#include "index/PatternCreator.h"
+#include "index/Permutation.h"
+#include "index/StxxlSortFunctors.h"
+#include "index/TextMetaData.h"
+#include "index/Vocabulary.h"
+#include "index/VocabularyMerger.h"
+#include "parser/ContextFileParser.h"
+#include "parser/TripleComponent.h"
+#include "parser/TurtleParser.h"
+#include "util/BufferedVector.h"
 #include "util/CancellationHandle.h"
+#include "util/File.h"
+#include "util/Forward.h"
+#include "util/HashMap.h"
 #include "util/MemorySize/MemorySize.h"
+#include "util/MmapVector.h"
+#include "util/json.h"
 
 using ad_utility::BufferedVector;
 using ad_utility::MmapVector;
