+++ src/index/IndexMetaDataImpl.h
@@ -5,12 +5,13 @@
 
 #pragma once
 
-#include "../util/File.h"
-#include "../util/Serializer/FileSerializer.h"
-#include "../util/Serializer/SerializeHashMap.h"
-#include "../util/Serializer/SerializeString.h"
-#include "./IndexMetaData.h"
-#include "./MetaDataHandler.h"
+#include "index/IndexMetaData.h"
+#include "index/MetaDataHandler.h"
+#include "util/File.h"
+#include "util/Serializer/ByteBufferSerializer.h"
+#include "util/Serializer/FileSerializer.h"
+#include "util/Serializer/SerializeHashMap.h"
+#include "util/Serializer/SerializeString.h"
 
 // _____________________________________________________________________________
 template <class MapType>
