+++ src/parser/ParsedQuery.cpp
@@ -4,17 +4,19 @@
 
 #include "ParsedQuery.h"
 
+#include <absl/strings/str_join.h>
+#include <absl/strings/str_split.h>
+
 #include <optional>
 #include <sstream>
 #include <string>
 #include <utility>
 #include <vector>
 
-#include "absl/strings/str_join.h"
-#include "absl/strings/str_split.h"
 #include "engine/sparqlExpressions/SparqlExpressionPimpl.h"
 #include "parser/RdfEscaping.h"
 #include "util/Conversions.h"
+#include "util/TransparentFunctors.h"
 
 using std::string;
 using std::vector;
