+++ src/util/ConstexprSmallString.h
@@ -5,8 +5,6 @@
 #pragma once
 
 #include <array>
-#include <cstring>
-#include <exception>
 #include <stdexcept>
 #include <string>
 #include <string_view>
